Complete refactor of the code
Most of the deletion and addition is because I have delete one package layer and moved it's item to upper layer package.

In this commit added a development mongoDB database so that the data can be dynamic and as per request.
In this commit RatingDTO is also removed as we are now not considering additional parameters for the request and response.

Due to it's flexibility, MongoDB is preferred as MongoDB repository have no schema requirement making it flexible to different record content.
Concept of primary key is not in the MongoDB database that's the major reason RatingDTO is now of no use as we can now directly can use TourRating object for request and response.